# MediathekArrApi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MediathekArrApi | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Media": {
      "get": {
        "tags": [
          "Media"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Media"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Media"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Media"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Rulesets": {
      "get": {
        "tags": [
          "Rulesets"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ruleset"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ruleset"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ruleset"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rulesets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Rulesets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Rulesets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Ruleset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Series": {
      "get": {
        "tags": [
          "Series"
        ],
        "parameters": [
          {
            "name": "tvdbId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Series"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Series"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Series"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Episode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "aired": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "runtime": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "seasonNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "episodeNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "MatchingStrategy": {
        "enum": [
          "SeasonAndEpisodeNumber",
          "ByAbsoluteEpisodeNumber",
          "ItemTitleIncludes",
          "ItemTitleExact",
          "ItemTitleEqualsAirdate"
        ]
      },
      "Media": {
        "type": "object",
        "properties": {
          "media_id": {
            "type": "integer",
            "format": "int32"
          },
          "media_name": {
            "type": "string"
          },
          "media_type": {
            "type": "string"
          },
          "media_tmdbId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "media_imdbId": {
            "type": "string",
            "nullable": true
          },
          "media_tvdbId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "Ruleset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mediaId": {
            "type": "integer",
            "format": "int32"
          },
          "topic": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "filters": {
            "type": "string"
          },
          "titleRegexRules": {
            "type": "string"
          },
          "episodeRegex": {
            "type": "string",
            "nullable": true
          },
          "seasonRegex": {
            "type": "string",
            "nullable": true
          },
          "matchingStrategy": {
            "$ref": "#/components/schemas/MatchingStrategy"
          },
          "media": {
            "$ref": "#/components/schemas/Media"
          }
        }
      },
      "Series": {
        "type": "object",
        "properties": {
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "germanName": {
            "type": "string",
            "nullable": true
          },
          "aliases": {
            "type": "string",
            "nullable": true
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "nextAired": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastAired": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cacheExpiry": {
            "type": "string",
            "format": "date-time"
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Episode"
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Media"
    },
    {
      "name": "Rulesets"
    },
    {
      "name": "Series"
    }
  ]
}